add tests for all process steps
