+++ completion-plugin/src/org/jb/cce/actions/BuildMultipleEvaluationsConfig.kt
@@ -3,7 +3,7 @@ package org.jb.cce.actions
 import org.jb.cce.Config
 import org.jb.cce.EvaluationWorkspace
 
-internal fun List<EvaluationWorkspace>.buildMultipleEvaluationsConfig(): Config {
+fun List<EvaluationWorkspace>.buildMultipleEvaluationsConfig(): Config {
     val existingConfig = this.first().readConfig()
     return Config.build(existingConfig.projectPath, existingConfig.language) {
         for (workspace in this@buildMultipleEvaluationsConfig) {
